Add pen-like behaviour
Add a pen-like behaviour to the grid by allowing the user to draw
by holding down the mouse button.
The `mousedown` and `mouseup` events are used to activate and deactivate
the pen, respectively.
Additionally, a fix is applied to the clearGrid function to ensure that
the pen-like behaviour is disabled after clicking the 'CLEAR' button.